RFC: x86/shstk: Add ARCH_SHSTK_SUPPRESS_UD
Add a quick POC for suppressing #UDs of shadow stack instructions. Only
support INCSSP because this is a POC. To enable the behavior:

arch_prctl(ARCH_SHSTK_ENABLE, ARCH_SHSTK_SUPPRESS_UD);

It is per-thread like the rest of them. When enabled, if an INCSSP
triggers a #UD, it will be skipped over. There is no emulation, it just
jumps to the next instruction.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>