RFC: x86/shstk: Add ARCH_SHSTK_SUPPRESS_CP
Add a quick POC for suppressing #CPs caused by user shadow stack
violations. Only support near-ret for now, because this is a POC. To
enable the behavior:

arch_prctl(ARCH_SHSTK_ENABLE, ARCH_SHSTK_SUPPRESS_CP);

It is per-thread like the rest of them. When enabled, the kernel will
emulate the RET as if shadow stack is not enabled. So there will be no
shadow stack POP and the SSP will be not be adjusted.

Signed-off-by: Rick Edgecombe <rick.p.edgecombe@intel.com>